Added graceTime, minShowTime and taskInProgress.
graceTime replaces the old delay property. It uses a NSTimer for greater efficiency. If set, it prevents the HUD from showing at all if the task does not run at least graceTime seconds.
minShowTime is similarly implemented using a NSTimer. If set, it denotes the minimum time the HUD will be shown (even if the task finishes prematurely).
taskInProgress can be used to notify the HUD of the task execution progress when externally managing the task's thread (needed for grace time)