need visibility to be able to use this for certain displays
discovered that HUDs were being displayed sideways or upside-down when
using the .view.window of a popped UINavigationController as the root
for a HUD view when returning.  if using the current view, the HUD would
not appear.  by using this call under that circumstance, the problem was
solved … so made it public.