Made label into a property
With label as a private ivar, I couldn’t find a way to extend the class and add functionality (in my case, I needed to make the label three lines long for a certain message). If we make it into a property, I can at least declare label as @NSManaged and refer to it in my subclass. (Without violating encapsulation and keeping it private.)