Remove from superview even if animation wasn't finished
Steps to reproduce in the sample app
1. Tap on a "Determinate mode" to bring the HUD
2. While the HUS is still presented background the app
3. Wait for 5 seconds until the HUD should have disappered
4. Foreground the app

Expected: no HUD, you can interact with the app.
Result without this: you can't interact with the app.

It happens because HUD wasn't removed from its superview. And it wasn't
removed because the app was in background when `hideAnimated:` was
called and animation wasn't finished.